Split out `ControlledPromise`.
